pyproject.toml
[tool.poetry]
name = "modernnomads"
version = "0.1.0"
description = ""
authors = ["lemonandlimee <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
